Removing Authentication functionality for now.
