+++ src/main/java/uk/gov/companieshouse/accounts/filing/configuration/ApplicationConfiguration.java
@@ -1,0 +1,25 @@
+package uk.gov.companieshouse.accounts.filing.configuration;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import uk.gov.companieshouse.logging.Logger;
+import uk.gov.companieshouse.logging.LoggerFactory;
+
+@Configuration
+public class ApplicationConfiguration {
+
+    @Value("${application.namespace}")
+    private String applicationNameSpace;
+
+    /**
+     * Creates the logger used by the application.
+     *
+     * @return the logger
+     */
+    @Bean
+    public Logger logger() {
+        return LoggerFactory.getLogger(applicationNameSpace);
+    }
+
+}
