+++ src/main/java/uk/gov/companieshouse/accounts/filing/configuration/WebMvcConfig.java
@@ -1,26 +1,22 @@
-package uk.gov.companieshouse.account.filing.configuration;
+package uk.gov.companieshouse.accounts.filing.configuration;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import uk.gov.companieshouse.account.filing.security.LoggingInterceptor;
-import uk.gov.companieshouse.account.filing.security.UserAuthenticationInterceptor;
+import uk.gov.companieshouse.accounts.filing.security.LoggingInterceptor;
 
 @Component
 public class WebMvcConfig implements WebMvcConfigurer {
     private final LoggingInterceptor loggingInterceptor;
-    private final UserAuthenticationInterceptor userAuthenticationInterceptor;
 
     @Autowired
-    public WebMvcConfig(LoggingInterceptor loggingInterceptor, UserAuthenticationInterceptor userAuthenticationInterceptor) {
+    public WebMvcConfig(LoggingInterceptor loggingInterceptor) {
         this.loggingInterceptor = loggingInterceptor;
-        this.userAuthenticationInterceptor = userAuthenticationInterceptor;
     }
 
     @Override
     public void addInterceptors(final InterceptorRegistry registry) {
         registry.addInterceptor(loggingInterceptor);
-        registry.addInterceptor(userAuthenticationInterceptor);
     }
 }
