+++ src/main/java/uk/gov/companieshouse/accounts/filing/controller/AccountsFilingController.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.account.filing.controller;
+package uk.gov.companieshouse.accounts.filing.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -6,12 +6,12 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/account-filing/filing")
-public class AccountFilingController {
+@RequestMapping("/accounts-filing")
+public class AccountsFilingController {
+
     @GetMapping("/check")
-    ResponseEntity<?> checking() {
-        return ResponseEntity.ok().body("Testing complete");
+    public ResponseEntity<?> checking(){
+        return ResponseEntity.ok().body("it is just for testing");
+
     }
-
 }
-
