+++ src/main/java/uk/gov/companieshouse/accounts/filing/security/LoggingInterceptor.java
@@ -1,0 +1,33 @@
+package uk.gov.companieshouse.accounts.filing.security;
+
+import org.springframework.lang.NonNull;
+import org.springframework.stereotype.Component;
+import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.servlet.handler.HandlerInterceptorAdapter;
+import uk.gov.companieshouse.logging.Logger;
+import uk.gov.companieshouse.logging.util.RequestLogger;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+@Component
+public class LoggingInterceptor extends HandlerInterceptorAdapter implements RequestLogger {
+    private final Logger logger = null;
+
+
+@Override
+    public boolean preHandle(@NonNull HttpServletRequest request,
+                             @NonNull HttpServletResponse response,
+                             @NonNull Object handler) {
+        logStartRequestProcessing(request, logger);
+        return true;
+    }
+
+    @Override
+    public void postHandle(@NonNull HttpServletRequest request,
+                           @NonNull HttpServletResponse response,
+                           @NonNull Object handler,
+                           ModelAndView modelAndView) {
+        logEndRequestProcessing(request, response, logger);
+    }
+}
