+++ src/main/java/uk/gov/companieshouse/accounts/filing/security/WebSecurityConfig.java
@@ -1,6 +1,5 @@
-package uk.gov.companieshouse.account.filing.security;
+package uk.gov.companieshouse.accounts.filing.security;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
@@ -8,22 +7,11 @@ import org.springframework.security.config.annotation.web.builders.WebSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.config.http.SessionCreationPolicy;
-import uk.gov.companieshouse.logging.Logger;
 
 @Configuration
 @EnableWebSecurity
 @EnableGlobalMethodSecurity(prePostEnabled = true)
 public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
-
-    private final UserAuthenticationInterceptor userAuthenticationInterceptor;
-    private final Logger logger;
-
-    @Autowired
-    public WebSecurityConfig(UserAuthenticationInterceptor userAuthenticationInterceptor, Logger logger) {
-        this.userAuthenticationInterceptor = userAuthenticationInterceptor;
-        this.logger = logger;
-    }
-
     /**
      * Configure Http Security.
      */
