+++ src/main/resources/application.properties
@@ -1,1 +1,1 @@
-
+application.namespace="accounts-filing-api"
