--- src/test/java/uk/gov/companieshouse/account/filing/AccountFilingApiApplicationTests.java
@@ -1,13 +1,0 @@
-package uk.gov.companieshouse.account.filing;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AccountFilingApiApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
