+++ src/test/java/uk/gov/companieshouse/accounts/filing/AccountsFilingApiApplicationTests.java
@@ -1,0 +1,9 @@
+package uk.gov.companieshouse.accounts.filing;
+
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AccountsFilingApiApplicationTests {
+
+
+}
